refactor: split root reducer into subject matter specific reducers
