+++ src/client/redux/Auth/AuthActionTypes.js
@@ -1,0 +1,8 @@
+import keymirror from 'mirrarray';
+
+const AuthActionTypes = keymirror([
+    'LOGIN_SUCCESS',
+    'LOGOUT',
+]);
+
+export default AuthActionTypes;
