+++ src/client/redux/Auth/AuthActions.js
@@ -1,0 +1,10 @@
+import ActionTypes from './AuthActionTypes';
+
+export const loginSuccess = ({ username, userid }) => ({
+    type: ActionTypes.LOGIN_SUCCESS,
+    payload: { username, userid },
+});
+
+export const logout = {
+    type: ActionTypes.LOGOUT,
+};
