+++ src/client/redux/FileSystem/FileSystemActionTypes.js
@@ -1,0 +1,8 @@
+import keymirror from 'mirrarray';
+
+const UIActionTypes = keymirror([
+    'SET_CURRENT_FILENAME',
+    'SET_USER_CODE',
+]);
+
+export default UIActionTypes;
