+++ src/client/redux/FileSystem/FileSystemActions.js
@@ -1,0 +1,6 @@
+import ActionTypes from './FileSystemActionTypes';
+
+export const setUserCode = code => ({
+    type: ActionTypes.SET_USER_CODE,
+    payload: code,
+});
