+++ src/client/redux/UI/UIActionTypes.js
@@ -1,20 +1,13 @@
 import keymirror from 'mirrarray';
 
-const ActionTypes = keymirror([
+const UIActionTypes = keymirror([
     'START_EXECUTION_ANIMATION',
     'STOP_EXECUTION_ANIMATION',
     'TOGGLE_AUTO_COMPLETE',
-    'START_DEBUG_MODE',
-    'STOP_DEBUG_MODE',
-    'SET_DEBUG_OUTPUT',
     'SET_AST',
     'SET_EDITOR_MODE',
     'SET_EDITOR_THEME',
-    'SET_EXECUTION_RESULTS',
     'RESTORE_DEFAULTS',
-    'LOGIN_SUCCESS',
-    'LOGOUT',
-    'SET_USER_CODE',
 ]);
 
-export default ActionTypes;
+export default UIActionTypes;
